requirements.txt
altgraph==0.17
async-generator==1.10
async-timeout==3.0.1
asyncio-extras==1.3.2
asyncpraw==7.3.0
asyncprawcore==2.2.0
attrs==21.2.0
Automat==20.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==3.0.4
comtypes==1.1.10
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
discord.py==1.7.2
ffmpeg-python==0.2.0
future==0.18.2
google==3.0.0
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.7.5
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
itemadapter==0.3.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.6.3
multidict==5.1.0
parsel==1.6.0
praw==7.3.0
prawcore==2.2.0
priority==1.3.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
PyNaCl==1.4.0
pyOpenSSL==20.0.1
queuelib==1.6.1
requests==2.25.1
rfc3986==1.5.0
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.2.1
spotipy==2.18.0
Twisted==21.2.0
typing-extensions==3.10.0.0
update-checker==0.18.0
urllib3==1.26.5
w3lib==1.22.0
websocket-client==1.1.0
yarl==1.6.3
youtube-dl==2021.6.6
zope.interface==5.4.0
